:recycle: : 의존성 분리 (#225)
