+++ Match-Api/src/main/java/com/example/matchapi/project/converter/ProjectConverter.java
@@ -2,7 +2,6 @@ package com.example.matchapi.project.converter;
 
 import com.example.matchapi.donation.dto.DonationRes;
 import com.example.matchapi.donation.helper.DonationHelper;
-import com.example.matchapi.order.helper.OrderHelper;
 import com.example.matchapi.project.dto.ProjectReq;
 import com.example.matchapi.project.dto.ProjectRes;
 import com.example.matchapi.project.helper.ProjectHelper;
@@ -11,13 +10,9 @@ import com.example.matchcommon.annotation.Converter;
 import com.example.matchdomain.donation.entity.*;
 import com.example.matchdomain.donation.entity.enums.HistoryStatus;
 import com.example.matchdomain.donation.entity.enums.RegularPayStatus;
-import com.example.matchdomain.donation.repository.RegularPaymentRepository;
-import com.example.matchdomain.project.dto.ProjectDto;
 import com.example.matchdomain.project.entity.*;
 import com.example.matchdomain.project.entity.enums.ImageRepresentStatus;
-import com.example.matchdomain.project.entity.enums.ReportReason;
 import com.example.matchdomain.project.repository.ProjectRepository;
-import com.example.matchdomain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 
 import java.time.LocalDateTime;
@@ -33,7 +28,6 @@ import static com.example.matchdomain.project.entity.enums.ProjectStatus.BEFORE_
 public class ProjectConverter {
     private final ProjectHelper projectHelper;
     private final DonationHelper donationHelper;
-    private final RegularPaymentRepository regularPaymentRepository;
     private static final String FIRST_TIME = "T00:00:00";
     private static final String LAST_TIME = "T23:59:59";
     public ProjectRes.ProjectDetail projectImgList(List<ProjectImage> projectImage) {
@@ -257,36 +251,6 @@ public class ProjectConverter {
                 .build();
     }
 
-    public ProjectRes.ProjectLists convertToProjectToDto(ProjectDto result) {
-        List<String> imgUrlList = new ArrayList<>();
-        List<RegularPayment> regularPayments = regularPaymentRepository.findByProjectIdAndRegularPayStatus(result.getId(), RegularPayStatus.PROCEEDING);
-
-        for(RegularPayment regularPayment : regularPayments){
-            imgUrlList.add(regularPayment.getUser().getProfileImgUrl());
-        }
-
-        return ProjectRes.ProjectLists
-                .builder()
-                .projectId(result.getId())
-                .imgUrl(result.getImgUrl())
-                .title(result.getProjectName())
-                .usages(result.getUsages())
-                .kind(result.getProjectKind().getName())
-                .like(result.getLike())
-                .userProfileImages(imgUrlList)
-                .totalDonationCnt(imgUrlList.size())
-                .build();
-    }
-
-    public ProjectComment convertToComment(Long id, Long projectId, String comment) {
-        return ProjectComment
-                .builder()
-                .userId(id)
-                .comment(comment)
-                .projectId(projectId)
-                .build();
-    }
-
     public DonationHistory convertToDonationHistory(Long projectId, HistoryStatus historyStatus) {
         return DonationHistory
                 .builder()
@@ -332,18 +296,6 @@ public class ProjectConverter {
                 }
         );
         return projectLists;
-    }
-
-    public ProjectRes.CommentList projectComment(User user, ProjectComment result) {
-        return ProjectRes.CommentList.builder()
-                .commentId(result.getId())
-                .comment(result.getComment())
-                .commentDate(result.getCreatedAt())
-                .nickname(user.getNickname())
-                .profileImgUrl(user.getProfileImgUrl())
-                .userId(result.getUserId())
-                .isMy(result.getUserId().equals(user.getId()))
-                .build();
     }
 
     public List<DonationRes.Tutorial> convertToTutorialDonation(List<Project> projects) {
