+++ Match-Api/src/main/java/com/example/matchapi/user/service/UserService.java
@@ -3,21 +3,16 @@ package com.example.matchapi.user.service;
 import com.example.matchapi.common.model.AlarmType;
 import com.example.matchapi.donation.service.DonationService;
 import com.example.matchapi.order.dto.OrderRes;
-import com.example.matchapi.project.converter.ProjectConverter;
 import com.example.matchapi.user.converter.UserConverter;
 import com.example.matchapi.user.dto.UserReq;
 import com.example.matchapi.user.dto.UserRes;
 import com.example.matchcommon.annotation.RedissonLock;
 import com.example.matchcommon.exception.BadRequestException;
 import com.example.matchdomain.common.model.Status;
-import com.example.matchdomain.donation.entity.RegularPayment;
-import com.example.matchdomain.donation.repository.RegularPaymentRepository;
-import com.example.matchdomain.project.repository.ProjectUserAttentionRepository;
 import com.example.matchdomain.user.adaptor.UserAdaptor;
 import com.example.matchdomain.user.entity.User;
 import com.example.matchdomain.user.entity.enums.Alarm;
 import com.example.matchdomain.user.exception.ModifyEmailCode;
-import com.example.matchdomain.user.repository.UserRepository;
 import com.example.matchinfrastructure.config.s3.S3UploadService;
 import com.example.matchinfrastructure.oauth.apple.service.AppleAuthService;
 import lombok.RequiredArgsConstructor;
@@ -25,7 +20,6 @@ import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.util.List;
 
 import static com.example.matchapi.common.model.AlarmType.EVENT;
 import static com.example.matchcommon.constants.MatchStatic.*;
